refactor: Extract useRedirectionURL hook from StoreLink
Easier to reuse, could be extracted to cozy-ui in the future
