+++ src/ducks/reimbursements/Reimbursements.jsx
@@ -11,7 +11,7 @@ import Figure from 'cozy-ui/transpiled/react/Figure'
 import styles from 'ducks/reimbursements/Reimbursements.styl'
 import Loading from 'components/Loading'
 import { KonnectorChip } from 'components/KonnectorChip'
-import { StoreLink } from 'components/StoreLink'
+import StoreLink from 'components/StoreLink'
 import { Section, SectionTitle, SectionSeparator } from 'components/Section'
 import withFilters from 'components/withFilters'
 import { getYear } from 'date-fns'
